Restore order of tests in #waitForSendDoneFor:..., and replicate in #waitForDataFor:...
I hadn't quite understood what the comment was referring to, but it makes sense on returning from semaphore wait to *immediately* check if what we're waiting for has happened, before checking the timeout--in a race, success should win over failure. And, if the connection has been closed, that is the more meaningful error compared to a timeout. This same structure can be used in #waitForDataFor:... and seems superior there as well.