Refactor: create a common superclass for inline and remove parameter refactoring
Previously, inline parameter was subclass of remove parameter.
This was ok since they share all the state and some of the behavior.
However, precondition checking was very awkward and because of it
developer had to implement `myConditions` since inline and remove
have very different preconditions.
Creating this common superclass enables us to remove `myConditions`
and only implement preconditions using `applicabilityPreconditions`.