Remove old backward compatibility code from FFILibrary
In the past we could declare platform library name for FII via methods such as #macModuleName. But now it should override other methods and we kept the old code for backward compatibility.

This compatibility has been kept for multiple versions of Pharo. We can probably drop it now.

This also reduces the violation to testThatThereAreNoSelectorsRemainingThatAreSentButNotImplemented

Fixes #16280