Update the Playground shortcut
This change updates the open playground shortcut. Until now it was cmd + o + w because the tool was called workspace in the past. We could not use the p because this was used by the monticello browser. But in Pharo 12 we removed the Monticello browser and we can now have a better shortcut cmd + o + p (Command Open Playground).

I also removed some references to the workspace to unify the image around the term "playground".

Fixes #3393