Simplify Calypso packaging
This PR removes the package Calypso-SystemPlugins-ClassScripts-Queries-Tests because it did not contains any tests but only examples. 

One of the two example class is useless and got removed. I moved the only example class remaining in Calypso-SystemPlugins-ClassScripts-Queries and I gave it a more explicit name to make sure developers understand it is examples.