Move group metadate of rules class side
This change moves the method #group of rules from the instance side to the class side.
This is a metadata that will not change for any instance of the rule so it makes sense to have it class side. It will be useful for Angel to implement some features. Currently, if we want the group of a rule we need to create a new instance and this takes some times because of the initializations of the rules. This change will make it uch faster to know the group of a rule for which we do not have an instance.

There is a temporary hack to deprecate the instance side group method that could still be here for rule existing outside pharo to let them move the group on the class side without breaking the tools that will use the new class side method