Fix calypso browsing of Ring models
It is possible to browse a ring model in Calypso and this is used to browse external code for example.

I would like to replace the MCSnapshotBrowser by this but thi feature is currently broken. Here I'm proposing fixes to not rely on image entities in Calypso code. 
In the next version of Calypso it would be nice to have tests for this. 

Once this PR is integrated I will be able to remove the use of the Snapshot browser (written in Morphic) from Iceberg and then from Pharo