Update FormSet to not recompute each scalling all the time
Having a scaling that is different of 1 can slow down a lot the image. For example, if we set a background in the image then the perf will be horrible because we will scale the Form of the background all the time and it takes time.

I'm proposing to save the forms we are scaling in the FormSet. Like this we only have to scale once.

Fixes #16884

I'd like a review from @Rinzwind