Catch CmCommandAborted in Calypso
Calypso was written with Commander 1. But we started to introduce Commander2 commands also in SystemCommands. 

Here is a PR so that error handling of Calypso manages both Commander1 and 2 errors. This will allow me to migrate some code from Commander 1 to Commander 2. This should be the first step to fix a dependency problem with 'Commander-Core' (this requires a second step after this one)