Remove usage of cached repository from file base repositories
All FileBasedRepositories in Monticello are using a MCCachedRepository to save all fetched packages in a package cache. 

This can be useful when we have a remote file based repository to avoid using the network too much, but this seems overkill for local repositories. I propose to push that down in the hierarchy to use this mecanism only on remote file based repositories.

This might also fix some issues with TonelRepositories that sometimes fails to save code with complex slots