Fix bug in protocol management of Traits
It happens currently that some methods copied from traits are not well categorized with their protocols.

This happens because we add the Traits methods in the class we are compiling before we copy the old protocols of the class.

I'm fixing the problem be copying the old protocols sooner and I'm adding a regression test that was failing before and is now passing.

This should fix the CI failures of https://github.com/pharo-spec/Spec/pull/1565 and https://github.com/pharo-spec/Spec/pull/1573