Move execution environments out of the Kernel
Execution environments are not used until we load SUnit in the image. 

I think it should not be in the Kernel. I hesitated to add a new package for it but it seems overkill. Instead I propose to package the whole thing in SUnit. If we push the concept in the future then we can move it somewhere else.