Remove references to SystemAnnouncer from Breakpoint tests
This PR removes SystemAnnouncer references from breakpoint tests. In order to do that I had to make announcers in SystemEnvironment configurable. For now they point to SystemAnnouncer unique instance so it should change nothing in the behavior.

The goal is still to kill the SystemAnnouncer singleton

Requires https://github.com/pharo-project/pharo/pull/16999 (I did not wanted to risk a conflict :) )