Remove unused sound support
Sound support does not have the primitive to make it work anymore. The settings don't even have default values sometimes but they are called with an #ifTrue: after so this means that this code cannot even work. 

If we want a sound support it would be better to use for example the binding to LibVLC or something else. 

Fixes #16942