Add a cache slot after message selectors.
This slot is to be used by the interpreter so it can optimize message lookups.
This change is missing some kind of "versioning"mechanism that can be used to signal which methods have cache slots and which do not.

Methods with cache slots should still work in previous VMs (they just see a big compiled method that has a lot of nils stored at its literals).