Refactor: move `checkTemporaries` to `prepareForExecution`
This code is doing more preparation then condition checking.
It is calculating temporaries and assignments.
Unfortunately, it is also doing two check validating the
if the selection is right. I'm moving it to preparation but will
come back to this to futher separate the concerns of precondition
checking and setup.