Fix: extract setUp method refactoring now correctly handles existing selectors
This commit migrates changes of extract method to extract setUp method.

Besides it fixes a bug in extract method that caused the test from commit to be skipped.
The bug is following: if refactoring finds existing selector with the equivalent body
it will not perform creation of `setUp` method, but will just remove the code from the
source method.
My thinking is that this is due to the fact that there was an assumption made that 
existing selector will be another setUp method (either from the super class or from that class)
and that only thing the refactoring should do is to remove the code so it is not duplicated.
To fix this I've added a check to still include setUp (and create the method)
if the existing selector is not another setUp method.
In these cases the existing selector will have the same code as setUp. Now this is something
that we can change in the future if this is not optimal. Maybe we can have setUp call the
existing selector and reduce duplication, just one solution.