Fix renaming protocol with no method
In calypso if you try to rename a protocol that has no method, this does not produce anything. 
This is because to do the update, it updates the protocol of all methods inside the protocol. But here we have no method. 

I updated the code to handle this case

Fixes #15387