Simplify progress bar theme management
The theme management of the progress bar is a little messy because it gives a progress bar parameter that is never used. And some methods have "style" in the name but they should return colors. 

Here is a pass to simplify this. I'll then do a next step to ensure that Spec is using those methods to theme the progress bars in Spec because this is not the case currently