Simplify the mecanism of Metacello
Metacello is storing all the arguments given by users in an OrderedCollection with a selector associated. Once we want to execute an action, we instantiate an executor and it will perform all those selectors. 

This makes the code a bit weird. I propose to simplify it by directly instantiating an executor and delegating all the configuration to it. 

This makes the code more straightforward.

I also updated a little the comment