Cleaning of DateAndTime
DateAndTime class>>clock has no setter so its value cannot be changed.
Most of the users of this methods just forward things to Time to call a method of the same name.

I propose to deprecate #clock and use Time instead. Also, there is a moment in the bootstrap when the clock is not initialized and causes crashes. This remove the need for the initialization