Group remote Metacello repositories
I tried to do a bigger change that does not pass the building of Pharo. In order to make it simpler to debug, I'm trying to divide the PR into smaller changes and here is one

Metacello has a lot of packages to declare remote repositories and Monticello even has some of its own.
A propose as a first step to group all the Metacello ones because we don't have the granularity to not load them together and it just takes useless space in the package list to have so many packages with 1 class each.