Add a temporary hack to allow Iceberg's tests to run
Recently a refactoring of the announcements in SUnit broke some CI like Iceberg's CI. The origin was found and esteban will fix it. 
I propose to add a workaround in the meantime because I was trying to fix problems on Iceberg's CI and this bug make it difficult. 

See https://github.com/pharo-vcs/iceberg/issues/1993 for more infos